File that calculates compound interest
For this project, I had to make a file that would be able to calculate and display the compound interest over time based off of user inputs of initial money, money per month, interest rate, and the length of time for the interest to grow. I made a clock class and made functions to display, calculate, and operate the program. Per the requirements, I used exceptions to handle most of the errors that may happen during the runtime of the program.